Allow Custom Port Usage
This will run on fresh installs and ask the user if they wish to change their web port.

Error checking is used to ensure it is a port between 1025 and 65565.

`-p` will now be used with the desired port number for that session only. It does not save over the webserver info stored in the database.